renamed test to match fn
